feat: implement high-precision arrangement algorithm for 2.5D SAT nesting
- Introduced a new module for packing visible model footprints on the build plate using convex-hull SAT checks.
- Prioritizes dense on-plate fill and applies anchor intent, with a fallback for aggressive fit-rate rescue.
- Incorporates research from various papers on irregular nesting, constructive heuristics, and multi-objective search.
- Added types for arrangement models, transforms, and hull cache entries.
- Implemented core functions for computing arrangement updates, including hull generation, SAT checks, and placement scoring.
