+++ src/app/page.tsx
@@ -44,7 +44,11 @@ import {
   isBoundsOutsideVolume,
   shouldUsePreciseBoundsForTransform,
 } from '@/utils/modelBounds';
-import { convexHull2d } from '@/supports/Rafts/Crenelated/geometry/convexHull2d';
+import {
+  computeHighPrecisionArrangeUpdates,
+  type HullCacheEntry,
+  type ArrangeModel as HighPrecisionArrangeModel,
+} from '@/features/scene/arrange/highPrecisionArrange';
 
 // Domain Features
 import { useSceneCollectionManager } from '@/features/scene/useSceneCollectionManager';
@@ -155,15 +159,7 @@ export default function Home() {
     scale: THREE.Vector3;
   } | null>(null);
   const dragDepthRef = React.useRef(0);
-  const arrangeHullFootprintCacheRef = React.useRef<Map<string, {
-    points: THREE.Vector2[];
-    halfW: number;
-    halfD: number;
-    localMinX: number;
-    localMaxX: number;
-    localMinY: number;
-    localMaxY: number;
-  }>>(new Map());
+  const arrangeHullFootprintCacheRef = React.useRef<Map<string, HullCacheEntry>>(new Map());
 
   React.useEffect(() => {
     if (arrangePrecisionMode !== 'high_precision') return;
@@ -1218,1005 +1214,17 @@ export default function Home() {
     await sleep(0);
 
     try {
-      const SAT_EPS_MM = 0.05;
-      const spacing = Math.max(0, arrangeSpacingMm);
-      // minSpacing is the SAT hull-to-hull gap enforcing the requested spacing.
-      const minSpacing = spacing + SAT_EPS_MM;
-      const PERF_COMPLEX_SCENE = visibleModels.length >= 30;
-      const MAX_CANDIDATE_NEIGHBORS = PERF_COMPLEX_SCENE ? 20 : 36;
-      const MAX_VERTEX_PAIR_CANDIDATES = PERF_COMPLEX_SCENE ? 180 : 420;
-      const MAX_CONTACT_CANDIDATES = PERF_COMPLEX_SCENE ? 360 : 900;
-      const MAX_LATTICE_CANDIDATES = PERF_COMPLEX_SCENE ? 320 : 900;
-      const MAX_EVALUATED_CANDIDATES = PERF_COMPLEX_SCENE ? 420 : 1200;
-      const MAX_CANDIDATE_BUFFER = PERF_COMPLEX_SCENE ? 560 : 1500;
-
-      const minX = scene.view3dSettings.originMode === 'front_left' ? 0 : -scene.view3dSettings.widthMm * 0.5;
-      const maxX = minX + scene.view3dSettings.widthMm;
-      const minY = scene.view3dSettings.originMode === 'front_left' ? 0 : -scene.view3dSettings.depthMm * 0.5;
-      const maxY = minY + scene.view3dSettings.depthMm;
-
-      const modelTransformById = new Map(
-        scene.models.map((model) => [model.id, getArrangeTransform(model)] as const),
-      );
-
-      const quant = (n: number) => Math.round(n * 1e4) / 1e4;
-
-      type HullData = {
-        points: THREE.Vector2[];
-        halfW: number;
-        halfD: number;
-        localMinX: number;
-        localMaxX: number;
-        localMinY: number;
-        localMaxY: number;
-      };
-
-      const getHullAtRotation = (model: (typeof visibleModels)[number], rotationZ: number): HullData => {
-        const t = modelTransformById.get(model.id) ?? model.transform;
-        const positionAttr = model.geometry.geometry.getAttribute('position') as THREE.BufferAttribute;
-        if (!positionAttr || positionAttr.count < 3) {
-          return {
-            points: [
-              new THREE.Vector2(-1, -1),
-              new THREE.Vector2(1, -1),
-              new THREE.Vector2(1, 1),
-              new THREE.Vector2(-1, 1),
-            ],
-            halfW: 1,
-            halfD: 1,
-            localMinX: -1,
-            localMaxX: 1,
-            localMinY: -1,
-            localMaxY: 1,
-          };
-        }
-
-        const key = [
-          model.geometry.geometry.uuid,
-          quant(t.rotation.x),
-          quant(t.rotation.y),
-          quant(rotationZ),
-          quant(t.scale.x),
-          quant(t.scale.y),
-          quant(t.scale.z),
-        ].join('|');
-
-        const cached = arrangeHullFootprintCacheRef.current.get(key);
-        if (cached) return cached;
-
-        const matrix = new THREE.Matrix4().compose(
-          new THREE.Vector3(0, 0, 0),
-          new THREE.Quaternion().setFromEuler(new THREE.Euler(
-            t.rotation.x,
-            t.rotation.y,
-            rotationZ,
-            t.rotation.order,
-          )),
-          t.scale,
-        );
-
-        const center = model.geometry.center;
-        // Single pass: collect strided sample AND guaranteed extremal vertices in 8 directions.
-        // Extremal coverage prevents the hull from underestimating the true geometry boundary.
-        const targetSamplesBase = 8000;
-        const stride = Math.max(1, Math.floor(positionAttr.count / targetSamplesBase));
-        const points2d: THREE.Vector2[] = [];
-        const tmp = new THREE.Vector3();
-        const nE = 8;
-        const eDx = [1, -1, 0, 0, 0.7071068, 0.7071068, -0.7071068, -0.7071068];
-        const eDy = [0, 0, 1, -1, 0.7071068, -0.7071068, 0.7071068, -0.7071068];
-        const eDot = new Float64Array(nE).fill(-Infinity);
-        const eXArr = new Float32Array(nE);
-        const eYArr = new Float32Array(nE);
-        for (let i = 0; i < positionAttr.count; i++) {
-          tmp.set(
-            positionAttr.getX(i) - center.x,
-            positionAttr.getY(i) - center.y,
-            positionAttr.getZ(i) - center.z,
-          ).applyMatrix4(matrix);
-          const tx = tmp.x;
-          const ty = tmp.y;
-          if (i % stride === 0) points2d.push(new THREE.Vector2(tx, ty));
-          for (let d = 0; d < nE; d++) {
-            const dot = tx * eDx[d] + ty * eDy[d];
-            if (dot > eDot[d]) { eDot[d] = dot; eXArr[d] = tx; eYArr[d] = ty; }
-          }
-        }
-        for (let d = 0; d < nE; d++) {
-          if (Number.isFinite(eXArr[d]) && Number.isFinite(eYArr[d])) {
-            points2d.push(new THREE.Vector2(eXArr[d], eYArr[d]));
-          }
-        }
-
-        const hull = convexHull2d(points2d);
-        const points = hull.length >= 3
-          ? hull
-          : [
-            new THREE.Vector2(-1, -1),
-            new THREE.Vector2(1, -1),
-            new THREE.Vector2(1, 1),
-            new THREE.Vector2(-1, 1),
-          ];
-
-        let localMinX = Infinity;
-        let localMaxX = -Infinity;
-        let localMinY = Infinity;
-        let localMaxY = -Infinity;
-        for (const p of points) {
-          localMinX = Math.min(localMinX, p.x);
-          localMaxX = Math.max(localMaxX, p.x);
-          localMinY = Math.min(localMinY, p.y);
-          localMaxY = Math.max(localMaxY, p.y);
-        }
-
-        if (!Number.isFinite(localMinX) || !Number.isFinite(localMaxX) || !Number.isFinite(localMinY) || !Number.isFinite(localMaxY)) {
-          localMinX = -1;
-          localMaxX = 1;
-          localMinY = -1;
-          localMaxY = 1;
-        }
-
-        const next: HullData = {
-          points,
-          halfW: Math.max(1, (localMaxX - localMinX) * 0.5),
-          halfD: Math.max(1, (localMaxY - localMinY) * 0.5),
-          localMinX,
-          localMaxX,
-          localMinY,
-          localMaxY,
-        };
-
-        arrangeHullFootprintCacheRef.current.set(key, next);
-        return next;
-      };
-
-      const axesFromPolygon = (poly: THREE.Vector2[]) => {
-        const axes: THREE.Vector2[] = [];
-        for (let i = 0; i < poly.length; i++) {
-          const a = poly[i];
-          const b = poly[(i + 1) % poly.length];
-          const edge = new THREE.Vector2(b.x - a.x, b.y - a.y);
-          if (edge.lengthSq() <= 1e-10) continue;
-          axes.push(new THREE.Vector2(-edge.y, edge.x).normalize());
-        }
-        return axes;
-      };
-
-      const projectPolygon = (poly: THREE.Vector2[], center: THREE.Vector2, axis: THREE.Vector2) => {
-        let min = Infinity;
-        let max = -Infinity;
-        for (const p of poly) {
-          const dot = (p.x + center.x) * axis.x + (p.y + center.y) * axis.y;
-          min = Math.min(min, dot);
-          max = Math.max(max, dot);
-        }
-        return { min, max };
-      };
-
-      const polygonsOverlapWithSpacing = (
-        polyA: THREE.Vector2[],
-        centerA: THREE.Vector2,
-        polyB: THREE.Vector2[],
-        centerB: THREE.Vector2,
-        minSpacing: number,
-      ) => {
-        const axes = [...axesFromPolygon(polyA), ...axesFromPolygon(polyB)];
-        for (const axis of axes) {
-          const pa = projectPolygon(polyA, centerA, axis);
-          const pb = projectPolygon(polyB, centerB, axis);
-          if ((pa.max + minSpacing) <= pb.min || (pb.max + minSpacing) <= pa.min) {
-            return false;
-          }
-        }
-        return true;
-      };
-
-      type CollisionProxy = {
-        center: THREE.Vector2;
-        hull: THREE.Vector2[];
-        halfW: number;
-        halfD: number;
-        localMinX: number;
-        localMaxX: number;
-        localMinY: number;
-        localMaxY: number;
-      };
-
-      type Placed = CollisionProxy & {
-        model: (typeof visibleModels)[number];
-        rotationZ: number;
-      };
-
-      const worldBoundsAt = (proxy: CollisionProxy, center: THREE.Vector2) => ({
-        minX: center.x + proxy.localMinX,
-        maxX: center.x + proxy.localMaxX,
-        minY: center.y + proxy.localMinY,
-        maxY: center.y + proxy.localMaxY,
-      });
-
-      const intersectsBroadphase = (
-        a: CollisionProxy,
-        centerA: THREE.Vector2,
-        b: CollisionProxy,
-        centerB: THREE.Vector2,
-        pad: number,
-      ) => {
-        const ba = worldBoundsAt(a, centerA);
-        const bb = worldBoundsAt(b, centerB);
-        if (ba.maxX + pad <= bb.minX) return false;
-        if (bb.maxX + pad <= ba.minX) return false;
-        if (ba.maxY + pad <= bb.minY) return false;
-        if (bb.maxY + pad <= ba.minY) return false;
-        return true;
-      };
-
-      const withinPlateAt = (proxy: CollisionProxy, center: THREE.Vector2) => {
-        const wb = worldBoundsAt(proxy, center);
-        return wb.minX >= minX && wb.maxX <= maxX && wb.minY >= minY && wb.maxY <= maxY;
-      };
-
-      const canPlaceAt = (candidate: CollisionProxy, center: THREE.Vector2, others: CollisionProxy[]) => {
-        if (!withinPlateAt(candidate, center)) return false;
-        for (const other of others) {
-          if (!intersectsBroadphase(candidate, center, other, other.center, minSpacing)) continue;
-          if (polygonsOverlapWithSpacing(candidate.hull, center, other.hull, other.center, minSpacing)) {
-            return false;
-          }
-        }
-        return true;
-      };
-
-      const finalAnchor = (() => {
-        if (arrangeAnchorMode === 'front_left') return new THREE.Vector2(minX, minY);
-        if (arrangeAnchorMode === 'front_right') return new THREE.Vector2(maxX, minY);
-        if (arrangeAnchorMode === 'back_left') return new THREE.Vector2(minX, maxY);
-        if (arrangeAnchorMode === 'back_right') return new THREE.Vector2(maxX, maxY);
-        return new THREE.Vector2((minX + maxX) * 0.5, (minY + maxY) * 0.5);
-      })();
-      // HP packing density objective: always fill from top-left first, then
-      // apply requested anchor as a whole-group post-transform.
-      const packingAnchor = new THREE.Vector2(minX, minY);
-
-      const targetIdSet = new Set(visibleModels.map((m) => m.id));
-      const blockers: CollisionProxy[] = scene.models
-        .filter((m) => m.visible && !targetIdSet.has(m.id))
-        .map((m) => {
-          const t = modelTransformById.get(m.id) ?? m.transform;
-          const h = getHullAtRotation(m, t.rotation.z);
-          return {
-            center: new THREE.Vector2(t.position.x, t.position.y),
-            hull: h.points,
-            halfW: h.halfW,
-            halfD: h.halfD,
-            localMinX: h.localMinX,
-            localMaxX: h.localMaxX,
-            localMinY: h.localMinY,
-            localMaxY: h.localMaxY,
-          };
-        });
-
-      const makeRotationOptions = (currentZ: number) => {
-        if (!arrangeAllowRotateOnZ) return [currentZ];
-        const options: number[] = [];
-        const seen = new Set<number>();
-        const rotationStepDeg = PERF_COMPLEX_SCENE ? 30 : 15;
-        const push = (angle: number) => {
-          const twoPi = Math.PI * 2;
-          let a = angle % twoPi;
-          if (a < 0) a += twoPi;
-          const k = Number(a.toFixed(5));
-          if (seen.has(k)) return;
-          seen.add(k);
-          options.push(a);
-        };
-
-        push(currentZ);
-        for (let deg = 0; deg < 360; deg += rotationStepDeg) push(THREE.MathUtils.degToRad(deg));
-        return options;
-      };
-
-      // ── Candidate generation helper ─────────────────────────────────────────
-      // For every edge of each placed polygon, project the candidate hull flush
-      // against that edge (at exactly minSpacing separation) to generate nestling
-      // candidates.  This lets irregular hulls fill concavities left by neighbours.
-      const buildContactCandidates = (
-        h: HullData,
-        pool: CollisionProxy[],
-        minCX: number,
-        maxCX: number,
-        minCY: number,
-        maxCY: number,
-        out: Array<{ x: number; y: number }>,
-        maxOut: number,
-      ) => {
-        if (maxOut <= 0) return;
-        for (const other of pool) {
-          if (out.length >= maxOut) break;
-          const otherPoly = other.hull;
-          for (let ei = 0; ei < otherPoly.length; ei++) {
-            if (out.length >= maxOut) break;
-            const vA = otherPoly[ei];
-            const vB = otherPoly[(ei + 1) % otherPoly.length];
-            const ex = vB.x - vA.x;
-            const ey = vB.y - vA.y;
-            const len = Math.sqrt(ex * ex + ey * ey);
-            if (len < 1e-8) continue;
-            const nx = -ey / len;
-            const ny = ex / len;
-
-            // Extreme of candidate hull in -normal direction
-            let minDot = Infinity;
-            for (const p of h.points) {
-              const d = p.x * nx + p.y * ny;
-              if (d < minDot) minDot = d;
-            }
-            // Extreme of other hull in +normal direction
-            let maxOther = -Infinity;
-            for (const p of otherPoly) {
-              const d = (p.x + other.center.x) * nx + (p.y + other.center.y) * ny;
-              if (d > maxOther) maxOther = d;
-            }
-
-            // For each vertex of the other hull, compute the candidate center that
-            // places the candidate hull flush against this edge at that vertex.
-            // candidate_center · n = (vO + other.center) · n + minSpacing - minDot
-            for (const vO of otherPoly) {
-              if (out.length >= maxOut) break;
-              const targetDot = (vO.x + other.center.x) * nx + (vO.y + other.center.y) * ny + minSpacing - minDot;
-              if (Math.abs(ny) > 0.1) {
-                for (const cx of [other.center.x, minCX, maxCX, packingAnchor.x]) {
-                  const cy = (targetDot - cx * nx) / ny;
-                  out.push({ x: Math.min(maxCX, Math.max(minCX, cx)), y: Math.min(maxCY, Math.max(minCY, cy)) });
-                  if (out.length >= maxOut) break;
-                }
-              }
-              if (Math.abs(nx) > 0.1) {
-                for (const cy of [other.center.y, minCY, maxCY, packingAnchor.y]) {
-                  const cx = (targetDot - cy * ny) / nx;
-                  out.push({ x: Math.min(maxCX, Math.max(minCX, cx)), y: Math.min(maxCY, Math.max(minCY, cy)) });
-                  if (out.length >= maxOut) break;
-                }
-              }
-            }
-            void maxOther; // used for context; direct offset not needed with vertex iteration
-          }
-        }
-      };
-
-      const allShareSameGeometry = visibleModels.length > 0
-        && visibleModels.every((m) => m.geometry.geometry.uuid === visibleModels[0].geometry.geometry.uuid);
-      const SAME_GEOMETRY_FAST_PATH = allShareSameGeometry && visibleModels.length >= 18;
-      const USE_CONTACT_CANDIDATES = !(SAME_GEOMETRY_FAST_PATH && PERF_COMPLEX_SCENE);
-      const ENABLE_MULTI_ORDERING_RETRY = visibleModels.length <= 32;
-
-      // ── Pre-select best rotation for identical-geometry batches ─────────────
-      // Run a fast BLF simulation at each angle and keep the angle that fits most.
-      const simPackAtAngle = (angle: number): number => {
-        const simPlaced: CollisionProxy[] = [...blockers];
-        let count = 0;
-        for (const model of visibleModels) {
-          const h = getHullAtRotation(model, angle);
-          const proxy: CollisionProxy = {
-            center: new THREE.Vector2(),
-            hull: h.points, halfW: h.halfW, halfD: h.halfD,
-            localMinX: h.localMinX, localMaxX: h.localMaxX,
-            localMinY: h.localMinY, localMaxY: h.localMaxY,
-          };
-          const minCX = minX - h.localMinX;
-          const maxCX = maxX - h.localMaxX;
-          const minCY = minY - h.localMinY;
-          const maxCY = maxY - h.localMaxY;
-          if (minCX > maxCX || minCY > maxCY) continue;
-
-          const pitchX = Math.max(1, (h.localMaxX - h.localMinX) + minSpacing);
-          const pitchY = Math.max(1, (h.localMaxY - h.localMinY) + minSpacing);
-          const colsX = Math.ceil((maxCX - minCX) / pitchX) + 1;
-          const rowsY = Math.ceil((maxCY - minCY) / pitchY) + 1;
-          const cands: Array<{ cx: number; cy: number; d: number }> = [];
-          const addSim = (x: number, y: number) => {
-            const cx = Math.min(maxCX, Math.max(minCX, x));
-            const cy = Math.min(maxCY, Math.max(minCY, y));
-            const dx = cx - packingAnchor.x; const dy = cy - packingAnchor.y;
-            cands.push({ cx, cy, d: dx * dx + dy * dy });
-          };
-          for (let ix = 0; ix <= colsX; ix++) {
-            for (let iy = 0; iy <= rowsY; iy++) addSim(minCX + ix * pitchX, minCY + iy * pitchY);
-          }
-          for (const other of simPlaced) {
-            const ob = worldBoundsAt(other, other.center);
-            addSim(ob.maxX + minSpacing - h.localMinX, other.center.y);
-            addSim(other.center.x, ob.maxY + minSpacing - h.localMinY);
-          }
-          cands.sort((a, b) => a.d - b.d);
-          for (const c of cands) {
-            const ctr = new THREE.Vector2(c.cx, c.cy);
-            if (canPlaceAt(proxy, ctr, simPlaced)) {
-              count++;
-              simPlaced.push({ ...proxy, center: ctr });
-              break;
-            }
-          }
-        }
-        return count;
-      };
-
-      const sharedBestRotation = (() => {
-        if (!allShareSameGeometry || visibleModels.length === 0) return null as number | null;
-        const probe = visibleModels[0];
-        const probeT = modelTransformById.get(probe.id) ?? probe.transform;
-        const rotOpts = makeRotationOptions(probeT.rotation.z);
-        if (rotOpts.length === 1) return rotOpts[0];
-        let bestAngle = rotOpts[0];
-        let bestCount = -1;
-        for (const angle of rotOpts) {
-          const n = simPackAtAngle(angle);
-          if (n > bestCount) { bestCount = n; bestAngle = angle; }
-        }
-        return bestAngle;
-      })();
-
-      // ── Model ordering ───────────────────────────────────────────────────────
-      const areaAtBestAngle = (model: (typeof visibleModels)[number]) => {
-        const angle = sharedBestRotation ?? (modelTransformById.get(model.id) ?? model.transform).rotation.z;
-        const h = getHullAtRotation(model, angle);
-        return Math.max(1, (h.localMaxX - h.localMinX) * (h.localMaxY - h.localMinY));
-      };
-      const modelOrder = [...visibleModels].sort((a, b) => areaAtBestAngle(b) - areaAtBestAngle(a));
-
-      // ── Placement attempt function ───────────────────────────────────────
-      // Greedy BLF is ordering-sensitive — different model orderings yield
-      // different packing configurations.  By retrying with alternative orderings
-      // when models spill, we eliminate the counter-intuitive case where smaller
-      // spacing fits fewer models than larger spacing.
-      const attemptPlacement = (order: typeof modelOrder) => {
-      const placed: Placed[] = [];
-      let spills: Placed[] = [];
-
-      // ── Per-model placement ──────────────────────────────────────────────────
-      for (const model of order) {
-        const t = modelTransformById.get(model.id) ?? model.transform;
-        const neighborPool: CollisionProxy[] = [...placed, ...blockers];
-        const candidateNeighbors = neighborPool.length > MAX_CANDIDATE_NEIGHBORS
-          ? neighborPool
-            .slice()
-            .sort((a, b) => a.center.distanceToSquared(packingAnchor) - b.center.distanceToSquared(packingAnchor))
-            .slice(0, MAX_CANDIDATE_NEIGHBORS)
-          : neighborPool;
-        const collisionPool = [...placed, ...blockers];
-
-        const angleOptions = (() => {
-          const base = makeRotationOptions(t.rotation.z);
-          if (SAME_GEOMETRY_FAST_PATH && sharedBestRotation != null) {
-            const alternatives = base.filter((a) => Math.abs(a - sharedBestRotation) > 1e-5).slice(0, 2);
-            return [sharedBestRotation, ...alternatives];
-          }
-          if (sharedBestRotation == null) return base;
-          const prioritized = [sharedBestRotation, ...base.filter((a) => Math.abs(a - sharedBestRotation) > 1e-5)];
-          return prioritized;
-        })();
-
-        let best: { proxy: CollisionProxy; center: THREE.Vector2; rotationZ: number; score: number } | null = null;
-
-        for (const rotationZ of angleOptions) {
-          const h = getHullAtRotation(model, rotationZ);
-          const candidateProxy: CollisionProxy = {
-            center: new THREE.Vector2(),
-            hull: h.points, halfW: h.halfW, halfD: h.halfD,
-            localMinX: h.localMinX, localMaxX: h.localMaxX,
-            localMinY: h.localMinY, localMaxY: h.localMaxY,
-          };
-
-          const minCenterX = minX - h.localMinX;
-          const maxCenterX = maxX - h.localMaxX;
-          const minCenterY = minY - h.localMinY;
-          const maxCenterY = maxY - h.localMaxY;
-          if (minCenterX > maxCenterX || minCenterY > maxCenterY) continue;
-
-          const seen = new Set<string>();
-          const cands: Array<{ x: number; y: number; sortKey: number }> = [];
-          const addCandidate = (x: number, y: number) => {
-            if (cands.length >= MAX_CANDIDATE_BUFFER) return;
-            const cx = Math.min(maxCenterX, Math.max(minCenterX, x));
-            const cy = Math.min(maxCenterY, Math.max(minCenterY, y));
-            const k = `${cx.toFixed(2)}:${cy.toFixed(2)}`;
-            if (seen.has(k)) return;
-            seen.add(k);
-            const dx = cx - packingAnchor.x; const dy = cy - packingAnchor.y;
-            cands.push({ x: cx, y: cy, sortKey: dx * dx + dy * dy });
-          };
-
-          // Plate corners / anchor seed
-          addCandidate(packingAnchor.x, packingAnchor.y);
-          addCandidate(minCenterX, minCenterY);
-          addCandidate(maxCenterX, minCenterY);
-          addCandidate(minCenterX, maxCenterY);
-          addCandidate(maxCenterX, maxCenterY);
-
-          // AABB touch positions — fast coverage for axis-aligned shapes
-          let vertexPairCount = 0;
-          for (const other of candidateNeighbors) {
-            if (cands.length >= MAX_CANDIDATE_BUFFER) break;
-            const ob = worldBoundsAt(other, other.center);
-            // Right of / left of / above / below neighbour
-            addCandidate(ob.maxX + minSpacing - h.localMinX, other.center.y);
-            addCandidate(ob.minX - minSpacing - h.localMaxX, other.center.y);
-            addCandidate(other.center.x, ob.maxY + minSpacing - h.localMinY);
-            addCandidate(other.center.x, ob.minY - minSpacing - h.localMaxY);
-            // Cross — align to packing anchor row/col
-            addCandidate(ob.maxX + minSpacing - h.localMinX, packingAnchor.y);
-            addCandidate(packingAnchor.x, ob.maxY + minSpacing - h.localMinY);
-            // Vertex-to-vertex nestling: place candidate hull so vertex vC aligns
-            // with placed vertex vO, offset by minSpacing along each axis.
-            for (const vO of other.hull) {
-              if (vertexPairCount >= MAX_VERTEX_PAIR_CANDIDATES) break;
-              for (const vC of h.points) {
-                if (vertexPairCount >= MAX_VERTEX_PAIR_CANDIDATES) break;
-                const wx = other.center.x + vO.x;
-                const wy = other.center.y + vO.y;
-                addCandidate(wx - vC.x + minSpacing, wy - vC.y);
-                addCandidate(wx - vC.x - minSpacing, wy - vC.y);
-                addCandidate(wx - vC.x, wy - vC.y + minSpacing);
-                addCandidate(wx - vC.x, wy - vC.y - minSpacing);
-                vertexPairCount++;
-              }
-            }
-          }
-
-          // SAT-contact edge candidates — places new hull flush against each placed edge
-          if (USE_CONTACT_CANDIDATES) {
-            const contactRaw: Array<{ x: number; y: number }> = [];
-            buildContactCandidates(
-              h,
-              candidateNeighbors,
-              minCenterX,
-              maxCenterX,
-              minCenterY,
-              maxCenterY,
-              contactRaw,
-              MAX_CONTACT_CANDIDATES,
-            );
-            for (const c of contactRaw) addCandidate(c.x, c.y);
-          }
-
-          // Dense regular lattice + hex offset — ensures we always match axis-aligned packing capacity
-          const pitchX = Math.max(1, (h.localMaxX - h.localMinX) + minSpacing);
-          const pitchY = Math.max(1, (h.localMaxY - h.localMinY) + minSpacing);
-          const colsX = Math.ceil((maxCenterX - minCenterX) / pitchX) + 1;
-          const rowsY = Math.ceil((maxCenterY - minCenterY) / pitchY) + 1;
-          const totalLatticeCells = (colsX + 1) * (rowsY + 1) * 1.5; // account for hex offset
-          const latticeStride = Math.max(1, Math.ceil(totalLatticeCells / Math.max(1, MAX_LATTICE_CANDIDATES)));
-          let latticeCounter = 0;
-          for (let ix = 0; ix <= colsX; ix++) {
-            for (let iy = 0; iy <= rowsY; iy++) {
-              if ((latticeCounter++ % latticeStride) !== 0) continue;
-              addCandidate(minCenterX + ix * pitchX, minCenterY + iy * pitchY);
-              // Hex-offset row: shift every other row by half pitch
-              if (iy % 2 === 1) {
-                addCandidate(minCenterX + ix * pitchX + pitchX * 0.5, minCenterY + iy * pitchY);
-              }
-            }
-          }
-
-          // Sort by anchor proximity and take first valid (BLF)
-          cands.sort((a, b) => a.sortKey - b.sortKey);
-          let localBest: { center: THREE.Vector2; anchorDistSq: number } | null = null;
-          for (let ci = 0; ci < cands.length && ci < MAX_EVALUATED_CANDIDATES; ci++) {
-            const c = cands[ci];
-            const center = new THREE.Vector2(c.x, c.y);
-            if (canPlaceAt(candidateProxy, center, collisionPool)) {
-              localBest = { center, anchorDistSq: c.sortKey };
-              break;
-            }
-          }
-
-          if (!localBest) continue;
-
-          if (!best || localBest.anchorDistSq < best.score) {
-            best = { proxy: candidateProxy, center: localBest.center, rotationZ, score: localBest.anchorDistSq };
-          }
-        }
-
-        if (best) {
-          placed.push({
-            model,
-            center: best.center,
-            rotationZ: best.rotationZ,
-            hull: best.proxy.hull,
-            halfW: best.proxy.halfW,
-            halfD: best.proxy.halfD,
-            localMinX: best.proxy.localMinX,
-            localMaxX: best.proxy.localMaxX,
-            localMinY: best.proxy.localMinY,
-            localMaxY: best.proxy.localMaxY,
-          });
-          continue;
-        }
-
-        const fallback = getHullAtRotation(model, t.rotation.z);
-        const outsideGap = Math.max(8, spacing);
-        const spillIndex = spills.length;
-        const spillCenter = new THREE.Vector2(
-          maxX + outsideGap - fallback.localMinX,
-          minY - fallback.localMinY + spillIndex * ((fallback.localMaxY - fallback.localMinY) + spacing),
-        );
-        spills.push({
-          model,
-          center: spillCenter,
-          rotationZ: t.rotation.z,
-          hull: fallback.points, halfW: fallback.halfW, halfD: fallback.halfD,
-          localMinX: fallback.localMinX, localMaxX: fallback.localMaxX,
-          localMinY: fallback.localMinY, localMaxY: fallback.localMaxY,
-        });
-      }
-
-      // Second chance pass: after initial fill, try to fit spills into gaps again
-      // using all rotation options while the plate context is now richer.
-      if (spills.length > 0) {
-        const remainingSpills: Placed[] = [];
-        const retryOrder = [...spills].sort((a, b) => {
-          const aw = a.localMaxX - a.localMinX;
-          const ad = a.localMaxY - a.localMinY;
-          const bw = b.localMaxX - b.localMinX;
-          const bd = b.localMaxY - b.localMinY;
-          return (aw * ad) - (bw * bd);
-        });
-
-        for (const spill of retryOrder) {
-          const model = spill.model;
-          const t = modelTransformById.get(model.id) ?? model.transform;
-          const neighborPool: CollisionProxy[] = [...placed, ...blockers];
-          const candidateNeighbors = neighborPool.length > MAX_CANDIDATE_NEIGHBORS
-            ? neighborPool
-              .slice()
-              .sort((a, b) => a.center.distanceToSquared(packingAnchor) - b.center.distanceToSquared(packingAnchor))
-              .slice(0, MAX_CANDIDATE_NEIGHBORS)
-            : neighborPool;
-          const collisionPool = [...placed, ...blockers];
-
-          const angleOptions = (() => {
-            const base = makeRotationOptions(t.rotation.z);
-            if (SAME_GEOMETRY_FAST_PATH && sharedBestRotation != null) {
-              const alternatives = base.filter((a) => Math.abs(a - sharedBestRotation) > 1e-5).slice(0, 2);
-              return [sharedBestRotation, ...alternatives];
-            }
-            if (sharedBestRotation == null) return base;
-            return [sharedBestRotation, ...base.filter((a) => Math.abs(a - sharedBestRotation) > 1e-5)];
-          })();
-
-          let best: { proxy: CollisionProxy; center: THREE.Vector2; rotationZ: number; score: number } | null = null;
-
-          for (const rotationZ of angleOptions) {
-            const h = getHullAtRotation(model, rotationZ);
-            const candidateProxy: CollisionProxy = {
-              center: new THREE.Vector2(),
-              hull: h.points, halfW: h.halfW, halfD: h.halfD,
-              localMinX: h.localMinX, localMaxX: h.localMaxX,
-              localMinY: h.localMinY, localMaxY: h.localMaxY,
-            };
-
-            const minCenterX = minX - h.localMinX;
-            const maxCenterX = maxX - h.localMaxX;
-            const minCenterY = minY - h.localMinY;
-            const maxCenterY = maxY - h.localMaxY;
-            if (minCenterX > maxCenterX || minCenterY > maxCenterY) continue;
-
-            const seen = new Set<string>();
-            const cands: Array<{ x: number; y: number; sortKey: number }> = [];
-            const addCandidate = (x: number, y: number) => {
-              if (cands.length >= MAX_CANDIDATE_BUFFER) return;
-              const cx = Math.min(maxCenterX, Math.max(minCenterX, x));
-              const cy = Math.min(maxCenterY, Math.max(minCenterY, y));
-              const k = `${cx.toFixed(2)}:${cy.toFixed(2)}`;
-              if (seen.has(k)) return;
-              seen.add(k);
-              const dx = cx - packingAnchor.x;
-              const dy = cy - packingAnchor.y;
-              cands.push({ x: cx, y: cy, sortKey: dx * dx + dy * dy });
-            };
-
-            addCandidate(packingAnchor.x, packingAnchor.y);
-            addCandidate(minCenterX, minCenterY);
-            addCandidate(maxCenterX, minCenterY);
-            addCandidate(minCenterX, maxCenterY);
-            addCandidate(maxCenterX, maxCenterY);
-
-            for (const other of candidateNeighbors) {
-              if (cands.length >= MAX_CANDIDATE_BUFFER) break;
-              const ob = worldBoundsAt(other, other.center);
-              addCandidate(ob.maxX + minSpacing - h.localMinX, other.center.y);
-              addCandidate(ob.minX - minSpacing - h.localMaxX, other.center.y);
-              addCandidate(other.center.x, ob.maxY + minSpacing - h.localMinY);
-              addCandidate(other.center.x, ob.minY - minSpacing - h.localMaxY);
-              addCandidate(ob.maxX + minSpacing - h.localMinX, packingAnchor.y);
-              addCandidate(packingAnchor.x, ob.maxY + minSpacing - h.localMinY);
-            }
-
-            if (USE_CONTACT_CANDIDATES) {
-              const contactRaw: Array<{ x: number; y: number }> = [];
-              buildContactCandidates(
-                h,
-                candidateNeighbors,
-                minCenterX,
-                maxCenterX,
-                minCenterY,
-                maxCenterY,
-                contactRaw,
-                MAX_CONTACT_CANDIDATES,
-              );
-              for (const c of contactRaw) addCandidate(c.x, c.y);
-            }
-
-            const pitchX = Math.max(1, (h.localMaxX - h.localMinX) + minSpacing);
-            const pitchY = Math.max(1, (h.localMaxY - h.localMinY) + minSpacing);
-            const colsX = Math.ceil((maxCenterX - minCenterX) / pitchX) + 1;
-            const rowsY = Math.ceil((maxCenterY - minCenterY) / pitchY) + 1;
-            const totalLatticeCells = (colsX + 1) * (rowsY + 1) * 1.5;
-            const latticeStride = Math.max(1, Math.ceil(totalLatticeCells / Math.max(1, MAX_LATTICE_CANDIDATES)));
-            let latticeCounter = 0;
-            for (let ix = 0; ix <= colsX; ix++) {
-              for (let iy = 0; iy <= rowsY; iy++) {
-                if ((latticeCounter++ % latticeStride) !== 0) continue;
-                addCandidate(minCenterX + ix * pitchX, minCenterY + iy * pitchY);
-                if (iy % 2 === 1) {
-                  addCandidate(minCenterX + ix * pitchX + pitchX * 0.5, minCenterY + iy * pitchY);
-                }
-              }
-            }
-
-            cands.sort((a, b) => a.sortKey - b.sortKey);
-            for (let ci = 0; ci < cands.length && ci < MAX_EVALUATED_CANDIDATES; ci++) {
-              const c = cands[ci];
-              const center = new THREE.Vector2(c.x, c.y);
-              if (canPlaceAt(candidateProxy, center, collisionPool)) {
-                if (!best || c.sortKey < best.score) {
-                  best = { proxy: candidateProxy, center, rotationZ, score: c.sortKey };
-                }
-                break;
-              }
-            }
-          }
-
-          if (best) {
-            placed.push({
-              model,
-              center: best.center,
-              rotationZ: best.rotationZ,
-              hull: best.proxy.hull,
-              halfW: best.proxy.halfW,
-              halfD: best.proxy.halfD,
-              localMinX: best.proxy.localMinX,
-              localMaxX: best.proxy.localMaxX,
-              localMinY: best.proxy.localMinY,
-              localMaxY: best.proxy.localMaxY,
-            });
-          } else {
-            remainingSpills.push(spill);
-          }
-        }
-
-        spills = remainingSpills;
-      }
-
-      return { placed, spills };
-      };
-
-      // ── Multi-ordering retry ─────────────────────────────────────────────────
-      // Try the primary ordering; if models spill, retry with alternative
-      // orderings and keep the result with fewest spills.
-      let bestResult = attemptPlacement(modelOrder);
-
-      if (ENABLE_MULTI_ORDERING_RETRY && bestResult.spills.length > 0) {
-        // Give spilled models first dibs on plate positions
-        const spillIds = new Set(bestResult.spills.map(s => s.model.id));
-        const spillFirstOrder = [
-          ...modelOrder.filter(m => spillIds.has(m.id)),
-          ...modelOrder.filter(m => !spillIds.has(m.id)),
-        ];
-        const attempt = attemptPlacement(spillFirstOrder);
-        if (attempt.spills.length < bestResult.spills.length) bestResult = attempt;
-      }
-
-      if (ENABLE_MULTI_ORDERING_RETRY && bestResult.spills.length > 0) {
-        // Smallest-first: small models fill gaps more flexibly
-        const attempt = attemptPlacement([...modelOrder].reverse());
-        if (attempt.spills.length < bestResult.spills.length) bestResult = attempt;
-      }
-
-      if (ENABLE_MULTI_ORDERING_RETRY && bestResult.spills.length > 0) {
-        // Interleaved: alternate large/small to balance coverage and gap-filling
-        const interleaved: typeof modelOrder = [];
-        let lo = 0; let hi = modelOrder.length - 1;
-        while (lo <= hi) {
-          interleaved.push(modelOrder[lo++]);
-          if (lo <= hi) interleaved.push(modelOrder[hi--]);
-        }
-        const attempt = attemptPlacement(interleaved);
-        if (attempt.spills.length < bestResult.spills.length) bestResult = attempt;
-      }
-
-      const placed = bestResult.placed;
-      const spills = bestResult.spills;
-
-      // ── Multi-axis compaction ────────────────────────────────────────────────
-      // Move each model toward the packing anchor (top-left), then independently squeeze it along
-      // pure X and pure Y.  Multiple passes let models ripple into the gaps
-      // opened by their neighbours.
-      const COMPACTION_PASSES = PERF_COMPLEX_SCENE ? 6 : 8;
-      const COMPACTION_STEPS = PERF_COMPLEX_SCENE ? 12 : 16;
-
-      const binarySlide = (entry: Placed, dir: THREE.Vector2, others: CollisionProxy[]) => {
-        const start = entry.center.clone();
-        if (dir.lengthSq() <= 1e-10) return false;
-        let lo = 0; let hi = 1;
-        for (let s = 0; s < COMPACTION_STEPS; s++) {
-          const mid = (lo + hi) * 0.5;
-          const c = new THREE.Vector2(start.x + dir.x * mid, start.y + dir.y * mid);
-          if (canPlaceAt(entry, c, others)) lo = mid; else hi = mid;
-        }
-        if (lo > 1e-4) {
-          entry.center.set(start.x + dir.x * lo, start.y + dir.y * lo);
-          return true;
-        }
-        return false;
-      };
-
-      for (let pass = 0; pass < COMPACTION_PASSES; pass++) {
-        let moved = false;
-        // Process farthest-from-anchor first so inner models don't block outer ones
-        const order = placed
-          .map((_, i) => i)
-          .sort((a, b) => placed[b].center.distanceToSquared(packingAnchor) - placed[a].center.distanceToSquared(packingAnchor));
-
-        for (const idx of order) {
-          const entry = placed[idx];
-          const others: CollisionProxy[] = [...blockers];
-          for (let oi = 0; oi < placed.length; oi++) {
-            if (oi !== idx) others.push(placed[oi]);
-          }
-          // Diagonal toward packing anchor
-          const toAnchor = new THREE.Vector2(packingAnchor.x - entry.center.x, packingAnchor.y - entry.center.y);
-          if (binarySlide(entry, toAnchor, others)) moved = true;
-          // Pure X toward packing anchor
-          const toAnchorX = new THREE.Vector2(packingAnchor.x - entry.center.x, 0);
-          if (binarySlide(entry, toAnchorX, others)) moved = true;
-          // Pure Y toward packing anchor
-          const toAnchorY = new THREE.Vector2(0, packingAnchor.y - entry.center.y);
-          if (binarySlide(entry, toAnchorY, others)) moved = true;
-        }
-
-        if (!moved) break;
-      }
-
-      // ── Post-pack anchor transform ───────────────────────────────────────────
-      // After dense top-left fill, translate the packed-on-plate group toward
-      // the user-selected anchor while preserving minimum spacing and blocker
-      // avoidance. Internal packed spacing stays unchanged under translation.
-      if (placed.length > 0) {
-        const getPlacedBounds = (entries: Placed[]) => {
-          let bMinX = Infinity;
-          let bMaxX = -Infinity;
-          let bMinY = Infinity;
-          let bMaxY = -Infinity;
-          for (const entry of entries) {
-            const wb = worldBoundsAt(entry, entry.center);
-            bMinX = Math.min(bMinX, wb.minX);
-            bMaxX = Math.max(bMaxX, wb.maxX);
-            bMinY = Math.min(bMinY, wb.minY);
-            bMaxY = Math.max(bMaxY, wb.maxY);
-          }
-          return { minX: bMinX, maxX: bMaxX, minY: bMinY, maxY: bMaxY };
-        };
-
-        const currentBounds = getPlacedBounds(placed);
-        const currentCenterX = (currentBounds.minX + currentBounds.maxX) * 0.5;
-        const currentCenterY = (currentBounds.minY + currentBounds.maxY) * 0.5;
-        const plateCenterX = (minX + maxX) * 0.5;
-        const plateCenterY = (minY + maxY) * 0.5;
-
-        let desiredDx = 0;
-        let desiredDy = 0;
-        if (arrangeAnchorMode === 'front_left') {
-          desiredDx = minX - currentBounds.minX;
-          desiredDy = minY - currentBounds.minY;
-        } else if (arrangeAnchorMode === 'front_right') {
-          desiredDx = maxX - currentBounds.maxX;
-          desiredDy = minY - currentBounds.minY;
-        } else if (arrangeAnchorMode === 'back_left') {
-          desiredDx = minX - currentBounds.minX;
-          desiredDy = maxY - currentBounds.maxY;
-        } else if (arrangeAnchorMode === 'back_right') {
-          desiredDx = maxX - currentBounds.maxX;
-          desiredDy = maxY - currentBounds.maxY;
-        } else {
-          desiredDx = plateCenterX - currentCenterX;
-          desiredDy = plateCenterY - currentCenterY;
-        }
-
-        const canShiftPlaced = (dx: number, dy: number) => {
-          for (const entry of placed) {
-            const shiftedCenter = new THREE.Vector2(entry.center.x + dx, entry.center.y + dy);
-            if (!withinPlateAt(entry, shiftedCenter)) return false;
-            for (const blocker of blockers) {
-              if (!intersectsBroadphase(entry, shiftedCenter, blocker, blocker.center, minSpacing)) continue;
-              if (polygonsOverlapWithSpacing(entry.hull, shiftedCenter, blocker.hull, blocker.center, minSpacing)) {
-                return false;
-              }
-            }
-          }
-          return true;
-        };
-
-        const maxDxLo = minX - currentBounds.minX;
-        const maxDxHi = maxX - currentBounds.maxX;
-        const maxDyLo = minY - currentBounds.minY;
-        const maxDyHi = maxY - currentBounds.maxY;
-        desiredDx = Math.min(maxDxHi, Math.max(maxDxLo, desiredDx));
-        desiredDy = Math.min(maxDyHi, Math.max(maxDyLo, desiredDy));
-
-        let appliedScale = 0;
-        if (canShiftPlaced(desiredDx, desiredDy)) {
-          appliedScale = 1;
-        } else {
-          let lo = 0;
-          let hi = 1;
-          for (let i = 0; i < 16; i++) {
-            const mid = (lo + hi) * 0.5;
-            if (canShiftPlaced(desiredDx * mid, desiredDy * mid)) {
-              lo = mid;
-            } else {
-              hi = mid;
-            }
-          }
-          appliedScale = lo;
-        }
-
-        if (appliedScale > 1e-4) {
-          const dx = desiredDx * appliedScale;
-          const dy = desiredDy * appliedScale;
-          for (const entry of placed) {
-            entry.center.set(entry.center.x + dx, entry.center.y + dy);
-          }
-        }
-      }
-
-      // ── Re-layout spills using regular column-based packing ──────────────
-      if (spills.length > 0) {
-        const outsideGap = Math.max(8, spacing);
-        let columnLeftX = maxX + outsideGap;
-        let columnYCursor = minY;
-        let columnMaxWidth = 0;
-
-        for (const entry of spills) {
-          const w = entry.localMaxX - entry.localMinX;
-          const d = entry.localMaxY - entry.localMinY;
-
-          // Wrap to next column if this model would exceed plate depth
-          if (columnYCursor > minY && (columnYCursor + d) > maxY) {
-            columnLeftX += columnMaxWidth + outsideGap;
-            columnMaxWidth = 0;
-            columnYCursor = minY;
-          }
-
-          entry.center.set(
-            columnLeftX - entry.localMinX,
-            columnYCursor - entry.localMinY,
-          );
-
-          columnYCursor += d + spacing;
-          columnMaxWidth = Math.max(columnMaxWidth, w);
-        }
-      }
-
-      const updates = [...placed, ...spills].map((entry) => {
-        const t = modelTransformById.get(entry.model.id) ?? entry.model.transform;
-        return {
-          id: entry.model.id,
-          transform: {
-            position: new THREE.Vector3(entry.center.x, entry.center.y, t.position.z),
-            rotation: new THREE.Euler(t.rotation.x, t.rotation.y, entry.rotationZ, t.rotation.order),
-            scale: t.scale.clone(),
-          },
-        };
+      const updates = computeHighPrecisionArrangeUpdates({
+        visibleModels: visibleModels as unknown as HighPrecisionArrangeModel[],
+        sceneModels: scene.models as unknown as HighPrecisionArrangeModel[],
+        widthMm: scene.view3dSettings.widthMm,
+        depthMm: scene.view3dSettings.depthMm,
+        originMode: scene.view3dSettings.originMode,
+        arrangeSpacingMm,
+        arrangeAllowRotateOnZ,
+        arrangeAnchorMode,
+        getArrangeTransform: (model) => getArrangeTransform(model as unknown as (typeof scene.models)[number]),
+        hullCache: arrangeHullFootprintCacheRef.current,
       });
 
       if (updates.length > 1) {
